refactor: extract responsibilities from AuthenticationService into dedicated services
Splits the 643-line AuthenticationService into focused modules:
- CredentialsService: credential DB operations (getByIdOrThrow, updateCounter)
- UsersService: user queries (getByIdOrThrow, getRoles)
- SessionsService: session validation with RLS context
- UserModerationService: throwIfBanned method added
- AuthenticationUtils: hashToken, transformCredentialForWebAuthn

No behavioral changes. Type check passes.
